delete_by_ids 级联删除

+++ codegen/__out__/rust/src/gen/base/tenant/tenant_dao.rs
@@ -2173,7 +2173,8 @@ pub async fn delete_by_ids(
       let domain_ids = old_model.domain_ids.clone();
       if !domain_ids.is_empty() {
         let mut args = QueryArgs::new();
-        let mut sql = "update base_tenant_domain set is_deleted=1 where".to_owned();
+        let mut sql = "update base_tenant_domain set is_deleted=1 where tenant_id=? and".to_owned();
+        args.push(id.clone().into());
         let arg = {
           let mut items = Vec::with_capacity(domain_ids.len());
           for item in domain_ids {
@@ -2199,7 +2200,8 @@ pub async fn delete_by_ids(
       let menu_ids = old_model.menu_ids.clone();
       if !menu_ids.is_empty() {
         let mut args = QueryArgs::new();
-        let mut sql = "update base_tenant_menu set is_deleted=1 where".to_owned();
+        let mut sql = "update base_tenant_menu set is_deleted=1 where tenant_id=? and".to_owned();
+        args.push(id.clone().into());
         let arg = {
           let mut items = Vec::with_capacity(menu_ids.len());
           for item in menu_ids {
@@ -2607,7 +2609,8 @@ pub async fn force_delete_by_ids(
       let domain_ids = old_model.domain_ids.clone();
       if !domain_ids.is_empty() {
         let mut args = QueryArgs::new();
-        let mut sql = "delete from base_tenant_domain where".to_owned();
+        let mut sql = "delete from base_tenant_domain where tenant_id=? and".to_owned();
+        args.push(id.clone().into());
         let mut items = Vec::with_capacity(domain_ids.len());
         for item in domain_ids {
           items.push("?");
@@ -2628,7 +2631,8 @@ pub async fn force_delete_by_ids(
       let menu_ids = old_model.menu_ids.clone();
       if !menu_ids.is_empty() {
         let mut args = QueryArgs::new();
-        let mut sql = "delete from base_tenant_menu where".to_owned();
+        let mut sql = "delete from base_tenant_menu where tenant_id=? and".to_owned();
+        args.push(id.clone().into());
         let mut items = Vec::with_capacity(menu_ids.len());
         for item in menu_ids {
           items.push("?");
